refactor: move world in its own package
